добавлено выпадение бонусов, обработка их сбора

+++ pacman/Assets/Scripts/Pacman/Data/DefData.cs
@@ -55,7 +55,7 @@ namespace Pacman.Data
 	{
 		public Dictionary<string, BonusTypeDef> types;
 		public List<BonusDropMomentDef> dropMoments;
-		public int accessTime;
+		public float accessTime;
 	}
 	
 	public class BonusTypeDef
